# chore(queue): don't allocate new message